refactor(slack): extract SlackClient into client module
Move the token-bearing HTTP client (SlackClient, retry/rate-limit handling, thread fetch + message resolution) into slack/client.rs, leaving slack.rs as the domain layer: error classification, URL parsing, wire-format structs, mention/message resolution, and YAML output. The client reaches domain items via descendant access, so no field-visibility changes were needed.

Both files are now <=400 lines (slack.rs 341, client.rs 375). Behavior unchanged: 449 tests (default) / 454 (js-rendering) pass, identical to the pre-split counts.
